+++ climpred/classes.py
@@ -785,7 +785,9 @@ class HindcastEnsemble(PredictionEnsemble):
     # ------------------
     # Analysis Functions
     # ------------------
-    def verify(self, name=None, metric='pearson_r', comparison='e2o', max_dof=False):
+    def verify(
+        self, name=None, metric='pearson_r', comparison='e2o', alignment='same_inits'
+    ):
         """Verifies the initialized ensemble against observations/verification data.
 
         This will automatically verify against all shared variables
@@ -799,8 +801,15 @@ class HindcastEnsemble(PredictionEnsemble):
                 observations/verification data. ('e2o' for ensemble mean to
                 observations/verification data. 'm2o' for each individual member to
                 observations/verification data).
-            max_dof (bool, default False): If ``True``, maximize the degrees of freedom
-                for each lag calculation.
+            alignment (str): which inits or verification times should be aligned?
+                - maximize/None: maximize the degrees of freedom by slicing ``hind`` and
+                ``verif`` to a common time frame at each lead.
+                - same_inits: slice to a common init frame prior to computing
+                metric. This philosophy follows the thought that each lead should be
+                based on the same set of initializations.
+                - same_verif: slice to a common/consistent verification time frame prior
+                to computing metric. This philosophy follows the thought that each lead
+                should be based on the same set of verification dates.
 
         Returns:
             Dataset of comparison results (if comparing to one observational product),
@@ -821,11 +830,11 @@ class HindcastEnsemble(PredictionEnsemble):
             input_dict=input_dict,
             metric=metric,
             comparison=comparison,
-            max_dof=max_dof,
+            alignment=alignment,
         )
 
     def compute_metric(
-        self, name=None, metric='pearson_r', comparison='e2o', max_dof=False
+        self, name=None, metric='pearson_r', comparison='e2o', alignment='same_inits'
     ):
         """Verifies the initialized ensemble against observations/verification data.
 
@@ -840,8 +849,15 @@ class HindcastEnsemble(PredictionEnsemble):
                 observations/verification data. ('e2o'
                 for ensemble mean to observations/verification data.
                 'm2o' for each individual member to observations/verification data).
-            max_dof (bool, default False): If ``True``, maximize the degrees of freedom
-                for each lag calculation.
+            alignment (str): which inits or verification times should be aligned?
+                - maximize/None: maximize the degrees of freedom by slicing ``hind`` and
+                ``verif`` to a common time frame at each lead.
+                - same_inits: slice to a common init frame prior to computing
+                metric. This philosophy follows the thought that each lead should be
+                based on the same set of initializations.
+                - same_verif: slice to a common/consistent verification time frame prior
+                to computing metric. This philosophy follows the thought that each lead
+                should be based on the same set of verification dates.
 
         Returns:
             Dataset of comparison results (if comparing to one observational product),
@@ -854,7 +870,7 @@ class HindcastEnsemble(PredictionEnsemble):
             PendingDeprecationWarning,
         )
         return self.verify(
-            name=name, metric=metric, comparison=comparison, max_dof=max_dof
+            name=name, metric=metric, comparison=comparison, alignment=alignment
         )
 
     def compute_uninitialized(self, name=None, metric='pearson_r', comparison='e2o'):
@@ -871,8 +887,15 @@ class HindcastEnsemble(PredictionEnsemble):
                 observations/verification data. ('e2o' for ensemble mean to
                 observations/verification data. 'm2o' for each individual member to
                 observations/verification data).
-            max_dof (bool, default False): If ``True``, maximize the degrees of freedom
-                for each lag calculation.
+            alignment (str): which inits or verification times should be aligned?
+                - maximize/None: maximize the degrees of freedom by slicing ``hind`` and
+                ``verif`` to a common time frame at each lead.
+                - same_inits: slice to a common init frame prior to computing
+                metric. This philosophy follows the thought that each lead should be
+                based on the same set of initializations.
+                - same_verif: slice to a common/consistent verification time frame prior
+                to computing metric. This philosophy follows the thought that each lead
+                should be based on the same set of verification dates.
 
         Returns:
             Dataset of comparison results (if comparing to one observational product),
@@ -897,7 +920,9 @@ class HindcastEnsemble(PredictionEnsemble):
             comparison=comparison,
         )
 
-    def compute_persistence(self, name=None, metric='pearson_r', max_dof=False):
+    def compute_persistence(
+        self, name=None, metric='pearson_r', alignment='same_inits'
+    ):
         """Verify against a persistence forecast of the observations/verification data.
 
         This simply applies some metric between the observational product and itself out
@@ -915,8 +940,15 @@ class HindcastEnsemble(PredictionEnsemble):
                 with which to compute the persistence forecast. If ``None``, compute
                 for all observations/verification data.
             metric (str, default 'pearson_r'): Metric to apply for verification.
-            max_dof (bool, default False): If ``True``, maximize the degrees of freedom
-                for each lag calculation.
+            alignment (str): which inits or verification times should be aligned?
+                - maximize/None: maximize the degrees of freedom by slicing ``hind`` and
+                ``verif`` to a common time frame at each lead.
+                - same_inits: slice to a common init frame prior to computing
+                metric. This philosophy follows the thought that each lead should be
+                based on the same set of initializations.
+                - same_verif: slice to a common/consistent verification time frame prior
+                to computing metric. This philosophy follows the thought that each lead
+                should be based on the same set of verification dates.
 
         Returns:
             Dataset of persistence forecast results (if ``name`` is not ``None``),
@@ -940,5 +972,8 @@ class HindcastEnsemble(PredictionEnsemble):
             'init': True,
         }
         return self._apply_climpred_function(
-            compute_persistence, input_dict=input_dict, metric=metric, max_dof=max_dof
+            compute_persistence,
+            input_dict=input_dict,
+            metric=metric,
+            alignment=alignment,
         )